set max-width v-container
