+++ front/assets/variables.scss
@@ -1,0 +1,9 @@
+@use 'vuetify/settings' with (
+  $container-max-widths: (
+    'sm': 600px,
+    'md': 720px,
+    'lg': 990px,
+    'xl': 990px,
+    'xxl': 990px
+  )
+);
