+++ front/nuxt.config.ts
@@ -1,14 +1,20 @@
 import { defineNuxtConfig } from 'nuxt/config'
+import vuetify from 'vite-plugin-vuetify'
 // https://v3.nuxtjs.org/api/configuration/nuxt.config
 export default defineNuxtConfig({
-  ssr: true,
   css: ['vuetify/styles'],
-  build: {
-    transpile: ["vuetify"],
-  },
   vite: {
-    define: {
-      "process.env.DEBUG": false,
+    ssr: {
+      noExternal: ['vuetify'],
     },
   },
+  modules: [
+    async (options, nuxt) => {
+      nuxt.hooks.hook('vite:extendConfig', config => config.plugins.push(
+        vuetify({
+          styles: { configFile: 'assets/variables.scss' }
+        })
+      ))
+    }
+  ],
 });
